functions/ml-train-model/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.23.1
google-cloud-storage==2.18.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
jinja2==3.1.6
db-dtypes==1.2.0
xgboost==2.1.1
lightgbm==4.5.0
